refactor: Address review comments
