+++ packages/client/src/appMapListItem.ts
@@ -1,0 +1,6 @@
+import { Metadata } from '@appland/models';
+
+export default interface AppMapListItem {
+  scenario_uuid: string;
+  metadata: Metadata;
+}
