+++ packages/client/src/findingStatusListItem.ts
@@ -1,0 +1,5 @@
+export default interface FindingStatusListItem {
+  app_id: string;
+  identity_hash: string;
+  status: string;
+}
