--- packages/client/src/types.d.ts
@@ -1,12 +1,0 @@
-import { Metadata } from '@appland/models';
-
-export type FindingStatusListItem = {
-  app_id: string;
-  identity_hash: string;
-  status: string;
-};
-
-export type AppMapListItem = {
-  scenario_uuid: string;
-  metadata: Metadata;
-};
