+++ packages/client/tsconfig.json
@@ -14,6 +14,6 @@
     "target": "es2017",
     "types": ["node", "jest"]
   },
-  "include": ["src/**/*", "bin/**/*"],
+  "include": ["src/**/*"],
   "exclude": ["**/node_modules"]
 }
